backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
torch>=2.2.0
transformers==4.48.2
numpy==2.2.2
pytest==8.3.4
requests==2.32.3
scipy==1.16.2
httpx==0.28.1
datasets==3.2.0
pandas==2.2.3
python-dotenv==1.0.1
